+++ src/commands/registerCommands.ts
@@ -1,4 +1,4 @@
-import { notationSign } from './images/notationSign';
+import { selectSigning } from './images/selectSignImage';
 import * as vscode from 'vscode';
 import { IRemoteTagTreeItem } from '../utils/DockerInstances';
 import { defaultSign } from './images/defaultSignImage';
@@ -9,6 +9,12 @@ export function registerCommands() {
     vscode.commands.registerCommand('securesupplychain.defaultKey', 
 	async (remoteTag: IRemoteTagTreeItem) => {
 		await defaultSign(remoteTag);
+		}
+	);
+
+	vscode.commands.registerCommand('securesupplychain.allKeys', 
+	async (remoteTag: IRemoteTagTreeItem) => {
+		await selectSigning(remoteTag);
 		}
 	);
 }
